Remove SCF and NSCF classes
